+++ src/main/java/com/ddang/usedauction/member/domain/Member.java
@@ -1,8 +1,11 @@
 package com.ddang.usedauction.member.domain;
 
 import com.ddang.usedauction.config.BaseTimeEntity;
+import com.ddang.usedauction.member.domain.enums.Role;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -22,36 +25,40 @@ import org.hibernate.annotations.SQLRestriction;
 @SQLRestriction("deleted_at IS NULL")
 public class Member extends BaseTimeEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
 
-    @Column(nullable = false)
-    private String memberId;
+  @Column(nullable = false)
+  private String memberId;
 
-    @Column(nullable = false)
-    private String passWord;
+  @Column(nullable = false)
+  private String passWord;
 
-    @Column(nullable = false)
-    private String email;
+  @Column(nullable = false)
+  private String email;
 
-    @Column
-    private boolean siteAlarm;
+  @Column
+  @Enumerated(EnumType.STRING)
+  private Role role;
 
-    @Column
-    private long point;
+  @Column
+  private boolean siteAlarm;
 
-    @Column
-    private String social;
+  @Column
+  private long point;
 
-    @Column
-    private String socialProviderId;
+  @Column
+  private String social;
 
-    @Column
-    private LocalDateTime deletedAt;
+  @Column
+  private String socialProviderId;
 
-    // 포인트 충전
-    public void addPoint(int point) {
-        this.point += point;
-    }
+  @Column
+  private LocalDateTime deletedAt;
+
+  // 포인트 충전
+  public void addPoint(int point) {
+    this.point += point;
+  }
 }
